Extract residuals from NMF models

Returns the residual matrix \(Y - \hat{Y}\) from a fitted NMF model. Requires the original observation matrix Y to be supplied.

For nmfre objects, residuals are computed from the BLUP reconstruction (\(Y - X(B_{blup})\)) by default. Set type = "fixed" to use fixed-effects only.

Usage

# S3 method for class 'nmfkc'
residuals(object, Y, ...)

# S3 method for class 'nmfae'
residuals(object, Y, ...)

# S3 method for class 'nmfre'
residuals(object, Y, type = c("blup", "fixed"), ...)

Arguments

object

A fitted model object.

Y

The original observation matrix used for fitting.

...

Not used.

type

For nmfre objects: "blup" (default) or "fixed".

Value

The residual matrix.

Examples

Y <- matrix(runif(50), 5, 10)
result <- nmfkc(Y, Q = 2)
#> Y(5,10)~X(5,2)B(2,10)...
#> 0sec
residuals(result, Y)
#>             [,1]         [,2]        [,3]        [,4]       [,5]          [,6]
#> [1,]  0.06877762  0.002189862  0.09631841 -0.00986392 -0.2359365  0.0891588849
#> [2,]  0.09005657 -0.050942563  0.30707456  0.09505555 -0.0853610  0.0754928911
#> [3,] -0.02157421  0.012694797 -0.20732145 -0.04220823 -0.1270709  0.0008282797
#> [4,]  0.01512540  0.369271568  0.09915286 -0.29021948  0.0804772  0.0190903959
#> [5,] -0.13766096 -0.385877201 -0.25980320  0.27469419  0.3250796 -0.1697080947
#>             [,7]        [,8]        [,9]       [,10]
#> [1,]  0.01527548  0.22333459 -0.12742870 -0.03844299
#> [2,] -0.10098869  0.15520933 -0.07483674 -0.29179450
#> [3,]  0.12270587  0.08143634 -0.02865040  0.23805077
#> [4,]  0.01668215 -0.25152138 -0.04836800 -0.10077568
#> [5,] -0.08723927 -0.14343032  0.26017105  0.16440962